feat(schema): simplify schema handling
BREAKING CHANGE: schema-path is now hardcoded to ${options.configRoot}/schema.json and can no longer be customized
Before this commit we allowed passing a custom schema path via options.schemaPath that allowed placing the schema in arbitrary locations. While this did provide a little more flexibility, it also required extra code and added complexity. In the name of KISS we are now always defaulting the schema path to ${options.configRoot}/schema.json, which we feel is a sane default and should satisfy the vast majority (if not all) use cases.