feat(base-config): allow defining an optional base-config
Environment-specific configs can now inherit values from an optional base config. This reduces a lot of duplication for config values that are the same across several environments.

Caveat: You can not put secrets into the base config because it's not possible to determine which encryption key should be used for them because base configs are env-independent